requirements.txt
Flask==2.2.5
git+https://github.com/openai/whisper.git
google-generativeai==0.7.2
python-dotenv==1.0.1
gTTS==2.5.2
